requirements.txt
gunicorn==19.1.1
Django==1.6.5
Fabric==1.9.0
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
PyYAML==3.11
South==0.8.4
amqp==1.4.6
amqplib==1.0.2
anyjson==0.3.3
billiard==3.3.0.18
binaryornot==0.3.0
celery==3.0.5
cookiecutter==0.8.0
django-filter==0.7
django-guardian==1.2.0
djangorestframework==2.3.13
djangorestframework-csv==1.3.3
ecdsa==0.11
kombu==2.3.2
mock==1.0.1
ordereddict==1.1
paramiko==1.14.0
pycrypto==2.6.1
pymongo==2.7.2
python-dateutil==1.5
python-memcached==1.48
pytz==2014.7
simplejson==3.6.5
six==1.7.2
mongoengine==0.8.7
django-rest-framework-mongoengine==2.0.2
wsgiref==0.1.2
https://github.com/ouinformatics/dockertask/zipball/master
requests
